# src/main/resources/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fitness Exercise API",
    "description": "API documentation for Fitness Exercise API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/refresh": {
      "post": {
        "tags": [
          "refresh-token-controller"
        ],
        "summary": "Refresh JWT Token",
        "description": "Refresh the JWT token using the provided refresh token",
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing refresh token or jwt token",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid JWT token or refresh token",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}